pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evolving-memory"
version = "0.1.0"
description = "Cognitive Trajectory Engine — topological memory graphs for LLM agents"
readme = "LICENSE"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0,<3.0",
    "networkx>=3.0,<4.0",
    "faiss-cpu>=1.7,<2.0",
    "sentence-transformers>=2.2,<4.0",
    "numpy>=1.24,<3.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0,<3.0"]
anthropic = ["anthropic>=0.30,<1.0"]
all = ["evolving-memory[openai,anthropic]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/evolving_memory"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
